Use `_` instead of `.` as separator for CMake prefixes
The underscore is just the established convention in CMake. Every good
guide, blog post, tutorial, and book uses it. Also, I found out that I
can add the underscore to the list of word separators in VS Code, or add
shortcuts for word part navigation, selection, and deletion. This
basically removes the only objective advantage of using dots.